Setup global types for requirements
