+++ src/requirements/types.ts
@@ -33,58 +33,6 @@ export type Course = {
   readonly acadGroup: string;
 };
 
-export type CourseTaken = {
-  readonly roster: string;
-  readonly courseId: number;
-  readonly code: string;
-  readonly subject: string;
-  readonly number: string;
-  readonly credits: number;
-};
-
-type RequirementCommon = {
-  /** Full name of the requirement. */
-  readonly name: string;
-  /** Description of the requirement. */
-  readonly description: string;
-  /** The source with more information on the requirement. (This should be a URL string.) */
-  readonly source: string;
-  /** If this is set to true, then an edge to the course doesn't count towards double counting. */
-  readonly allowCourseDoubleCounting?: true;
-};
-
-/**
- * @param T additional information only attached to credits and courses type.
- */
-type RequirementFulfillmentInformation<T = Record<string, unknown>> =
-  | {
-      readonly fulfilledBy: 'self-check';
-      // Currently unused.
-      readonly minCount?: number;
-    }
-  | ({
-      /** Defines how courses in a sub-requirement can be all counted towards a stat. */
-      readonly subRequirementProgress: 'every-course-needed' | 'any-can-count';
-      readonly fulfilledBy: 'credits' | 'courses';
-      /**
-       * The minimum count required to fulfill this requirement.
-       *
-       * - When fulfilledBy === 'credits', this field stores the min number of credits.
-       * - When fulfilledBy === 'courses', this field stores the min number of courses.
-       */
-      readonly minCount: number;
-    } & T)
-  | {
-      readonly fulfilledBy: 'toggleable';
-      readonly fulfillmentOptions: {
-        readonly [optionName: string]: {
-          readonly minCount: number;
-          readonly counting: 'credits' | 'courses';
-          readonly subRequirementProgress: 'every-course-needed' | 'any-can-count';
-          readonly description: string;
-        } & T;
-      };
-    };
 export type BaseRequirement = RequirementCommon & RequirementFulfillmentInformation;
 
 export type RequirementChecker = (course: Course) => boolean;
@@ -92,20 +40,6 @@ export type CollegeOrMajorRequirement = RequirementCommon &
   RequirementFulfillmentInformation<{
     readonly checker: RequirementChecker | readonly RequirementChecker[];
   }>;
-
-export type EligibleCourses = {
-  // "FA20": [123456, 42, 65536, /* and another crseId */]
-  readonly [semester: string]: readonly number[];
-};
-
-export type DecoratedCollegeOrMajorRequirement = RequirementCommon &
-  RequirementFulfillmentInformation<{ readonly courses: readonly EligibleCourses[] }>;
-
-export type RequirementWithIDSourceType = DecoratedCollegeOrMajorRequirement & {
-  readonly id: string;
-  readonly sourceType: 'College' | 'Major' | 'Minor';
-  readonly sourceSpecificName: string;
-};
 
 export type CollegeRequirements<R> = {
   readonly [collegeCode: string]: {
@@ -138,28 +72,9 @@ export type DecoratedRequirementsJson = {
   readonly minor: MajorRequirements<DecoratedCollegeOrMajorRequirement>;
 };
 
-export type RequirementFulfillmentStatistics = {
-  readonly fulfilledBy: 'courses' | 'credits' | 'self-check';
-  readonly minCountFulfilled: number;
-  readonly minCountRequired: number;
-};
-
-export type RequirementFulfillment = {
-  /** The original requirement object. */
-  readonly requirement: RequirementWithIDSourceType;
-  /** A list of courses that satisfy this requirement. */
-  readonly courses: readonly (readonly CourseTaken[])[];
-} & RequirementFulfillmentStatistics;
-
-export type GroupedRequirementFulfillmentReport = {
-  readonly groupName: 'College' | 'Major' | 'Minor';
-  readonly specific: string;
-  readonly reqs: readonly RequirementFulfillment[];
-};
-
 export type CrseInfo = {
   readonly roster: string;
-  crseIds: number[];
+  readonly crseIds: number[];
 };
 
 export type CompletedSubReqCourseSlot = {
